refactor(runtime): extract OTel span helpers to otel_spans module
Move start_conversation_context, start_llm_span, finish_llm_span,
start_tool_span, and serialize_history_for_span into a dedicated
otel_spans.rs module. These are pure functions with no dependency on
Orchestrator — they only need LLM types and OTel primitives.

orchestrator.rs shrinks by ~230 lines (2,690 -> ~2,460 remaining).
All 45 tests continue to pass.
