+++ crates/runtime/src/lib.rs
@@ -1,10 +1,12 @@
 pub mod bootstrap;
 pub mod metrics;
 pub mod orchestrator;
+pub(crate) mod otel_spans;
 pub mod scheduler;
 pub mod telemetry;
 
 pub use metrics::MetricsRecorder;
-pub use orchestrator::{start_conversation_context, Orchestrator, TurnResult};
+pub use orchestrator::{Orchestrator, TurnResult};
+pub use otel_spans::start_conversation_context;
 pub use scheduler::spawn_scheduler;
 pub use telemetry::{init_tracing, OtelGuard};
